API/requirements.txt
tensorflow==2.17.1 
fastapi
uvicorn
python-multipart
pillow
matplotlib
numpy  